Initial import of the S3 bucket and setting up the GH Action
This is syncing with a test bucket at the moment. When we are ready
to roll this out, we'll just have to change the `AWS_S3_BUCKET` to
the real S3 bucket.

We intentionally did not enable the `--delete` flag on the S3 sync
command for now, which would have deleted any files that are found
on the remote S3 bucket but not in the Git repo. The reason for this
is there are legacy files on the S3 bucket that we want to keep for
now, but don't want to add them to the Git repo as that would create
semi-permanent Git bloat.

This means deleting files on the Git repository _has no effect_ on
the remote S3 bucket (i.e. removing a file from here will not delete
it on S3).

If the need comes up in the future, we can scope the sync to only
the folders we care about and enable the `--delete` flag on them.

Co-authored-by: Jen Weber <j@jenweber.me>